+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ### Added
 
 - BC-609 - showing finished (formerly archived) tasks in new task overview
+- BC-921 - Adding toggle drag&drop feature for mobile devices on room overview
 
 ### Changed